+++ components/catalogs/front/src/components/ProductSelection/hooks/useInfiniteAttributeCriterionFactories.ts
@@ -27,15 +27,15 @@ type Result = {
 };
 
 const ALLOWED_ATTRIBUTE_TYPES = [
-    'identifier',
-    'text',
-    'textarea',
-    'simpleselect',
-    'multiselect',
-    'number',
-    'metric',
-    'boolean',
-    'date',
+    'pim_catalog_identifier',
+    'pim_catalog_text',
+    'pim_catalog_textarea',
+    'pim_catalog_simpleselect',
+    'pim_catalog_multiselect',
+    'pim_catalog_number',
+    'pim_catalog_metric',
+    'pim_catalog_boolean',
+    'pim_catalog_date',
 ];
 
 export const useInfiniteAttributeCriterionFactories = ({search = '', limit = 20}: QueryParams = {}): Result => {